Softmax files introduced. The softmax formula should be implemented for floats.

+++ spinnaker_graph_front_end/examples/TensorSample/tf_softmax_app_non_dynamic.py
@@ -20,7 +20,7 @@ tf.set_random_seed(0)
 # 2-D tensor `a`
 # [[1, 2, 3],
 #  [4, 5, 6]]
-k = tf.constant([1, 2, 3, 4, 5, 6], shape=[2, 3])
+k = tf.constant([1, 2, 3, 4, 5, 6], shape=[2, 3], dtype=tf.float32)
 
 c = tf.nn.softmax(k)
